Fix(rolling-upgrades base_version): Set base_version to 4.5
The "auto mode" base version doesn't work and it tries to install
base version that is the same as the upgraded version (e.g from
4.6.rc5 to 4.6.rc5).
Till I find where the exact issue in that code, switching the jobs
to use this workaround.
In this commit, adding htis workaround to all other upgrade jobs.